Move subtree/avail writing within TilesetBuilder
