+++ CDBTo3DTiles/src/CDBTo3DTiles.cpp
@@ -1,6 +1,5 @@
 #include "CDBTo3DTiles.h"
 #include "CDB.h"
-#include "FileUtil.h"
 #include "Gltf.h"
 #include "Math.h"
 #include "TileFormatIO.h"
@@ -106,463 +105,152 @@ void Converter::convert()
     std::map<std::string, std::vector<std::filesystem::path>> combinedTilesets;
     std::map<std::string, std::vector<Core::BoundingRegion>> combinedTilesetsRegions;
     std::map<std::string, Core::BoundingRegion> aggregateTilesetsRegion;
+    std::map<CDBDataset, std::filesystem::path> &datasetDirs = m_impl->datasetDirs;
+    m_impl->initializeImplicitTilingParameters();
 
-    if (m_impl->use3dTilesNext) {
-        int subtreeLevels = m_impl->subtreeLevels;
-        const uint64_t subtreeNodeCount = static_cast<int>((pow(4, subtreeLevels) - 1) / 3);
-        const uint64_t childSubtreeCount = static_cast<int>(pow(4, subtreeLevels)); // 4^N
+    cdb.forEachGeoCell([&](CDBGeoCell geoCell) {
+        m_impl->datasetCSSubtrees.clear();
+        datasetDirs.clear();
 
-        const uint64_t availabilityByteLength = static_cast<int>(
-            ceil(static_cast<double>(subtreeNodeCount) / 8.0));
-        const uint64_t nodeAvailabilityByteLengthWithPadding = alignTo8(availabilityByteLength);
-        const uint64_t childSubtreeAvailabilityByteLength = static_cast<int>(
-            ceil(static_cast<double>(childSubtreeCount) / 8.0));
-        const uint64_t childSubtreeAvailabilityByteLengthWithPadding = alignTo8(
-            childSubtreeAvailabilityByteLength);
-        m_impl->nodeAvailabilityByteLengthWithPadding = nodeAvailabilityByteLengthWithPadding;
-        m_impl->childSubtreeAvailabilityByteLengthWithPadding = childSubtreeAvailabilityByteLengthWithPadding;
+        // create directories for converted GeoCell
+        std::filesystem::path geoCellRelativePath = geoCell.getRelativePath();
+        std::filesystem::path geoCellAbsolutePath = m_impl->outputPath / geoCellRelativePath;
+        std::filesystem::path elevationDir = geoCellAbsolutePath / CDBTilesetBuilder::ELEVATIONS_PATH;
+        std::filesystem::path GTModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GTMODEL_PATH;
+        std::filesystem::path GSModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GSMODEL_PATH;
+        std::filesystem::path roadNetworkDir = geoCellAbsolutePath / CDBTilesetBuilder::ROAD_NETWORK_PATH;
+        std::filesystem::path railRoadNetworkDir = geoCellAbsolutePath
+                                                    / CDBTilesetBuilder::RAILROAD_NETWORK_PATH;
+        std::filesystem::path powerlineNetworkDir = geoCellAbsolutePath
+                                                    / CDBTilesetBuilder::POWERLINE_NETWORK_PATH;
+        std::filesystem::path hydrographyNetworkDir = geoCellAbsolutePath
+                                                        / CDBTilesetBuilder::HYDROGRAPHY_NETWORK_PATH;
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::Elevation, elevationDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSFeature, GSModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSModelGeometry, GSModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSModelTexture, GSModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTFeature, GTModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTModelGeometry_500, GTModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTModelTexture, GTModelDir));
+        datasetDirs.insert(
+            std::pair<CDBDataset, std::filesystem::path>(CDBDataset::RoadNetwork, roadNetworkDir));
+        datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::RailRoadNetwork,
+                                                                        railRoadNetworkDir));
+        datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::PowerlineNetwork,
+                                                                        powerlineNetworkDir));
+        datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::HydrographyNetwork,
+                                                                        hydrographyNetworkDir));
+                                          
 
-        const uint64_t headerByteLength = 24;
-
-        // dataset -> component selectors "CS1_CS2" -> subtree root "level_x_y -> subtree"
-        std::map<CDBDataset, std::map<std::string, std::map<std::string, SubtreeAvailability>>>
-            &datasetCSSubtrees = m_impl->datasetCSSubtrees;
-
-        std::vector<Core::BoundingRegion> boundingRegions;
-        std::vector<std::filesystem::path> tilesetJsonPaths;
-        cdb.forEachGeoCell([&](CDBGeoCell geoCell) {
-            datasetCSSubtrees.clear();
-
-            std::filesystem::path geoCellRelativePath = geoCell.getRelativePath();
-            std::filesystem::path geoCellAbsolutePath = m_impl->outputPath / geoCellRelativePath;
-            std::filesystem::path elevationDir = geoCellAbsolutePath / CDBTilesetBuilder::ELEVATIONS_PATH;
-            std::filesystem::path GSModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GSMODEL_PATH;
-            std::filesystem::path GTModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GTMODEL_PATH;
-            std::filesystem::path roadNetworkDir = geoCellAbsolutePath / CDBTilesetBuilder::ROAD_NETWORK_PATH;
-            std::filesystem::path railRoadNetworkDir = geoCellAbsolutePath
-                                                       / CDBTilesetBuilder::RAILROAD_NETWORK_PATH;
-            std::filesystem::path powerlineNetworkDir = geoCellAbsolutePath
-                                                        / CDBTilesetBuilder::POWERLINE_NETWORK_PATH;
-            std::filesystem::path hydrographyNetworkDir = geoCellAbsolutePath
-                                                          / CDBTilesetBuilder::HYDROGRAPHY_NETWORK_PATH;
-            std::map<CDBDataset, std::filesystem::path> datasetDirs;
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::Elevation, elevationDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSFeature, GSModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSModelGeometry, GSModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GSModelTexture, GSModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTFeature, GTModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTModelGeometry_500, GTModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::GTModelTexture, GTModelDir));
-            datasetDirs.insert(
-                std::pair<CDBDataset, std::filesystem::path>(CDBDataset::RoadNetwork, roadNetworkDir));
-            datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::RailRoadNetwork,
-                                                                            railRoadNetworkDir));
-            datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::PowerlineNetwork,
-                                                                            powerlineNetworkDir));
-            datasetDirs.insert(std::pair<CDBDataset, std::filesystem::path>(CDBDataset::HydrographyNetwork,
-                                                                            hydrographyNetworkDir));
-
-            cdb.forEachElevationTile(geoCell, [&](CDBElevation elevation) {
-                m_impl->addElevationToTilesetCollection(elevation, cdb, elevationDir);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->elevationTilesets);
-            std::unordered_map<CDBTile, Texture>().swap(m_impl->processedParentImagery);
-
-            cdb.forEachGSModelTile(geoCell, [&](CDBGSModels GSModel) {
-                m_impl->addGSModelToTilesetCollection(GSModel, GSModelDir);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->GSModelTilesets, false);
-
-            cdb.forEachGTModelTile(geoCell, [&](CDBGTModels GTModel) {
-                m_impl->addGTModelToTilesetCollection(GTModel, GTModelDir);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->GTModelTilesets);
-
-            cdb.forEachRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &roadNetwork) {
-                m_impl->addVectorToTilesetCollection(roadNetwork, roadNetworkDir, m_impl->roadNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->roadNetworkTilesets);
-
-            cdb.forEachRailRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &railRoadNetwork) {
-                m_impl->addVectorToTilesetCollection(railRoadNetwork,
-                                                     railRoadNetworkDir,
-                                                     m_impl->railRoadNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->railRoadNetworkTilesets);
-
-            cdb.forEachPowerlineNetworkTile(geoCell, [&](const CDBGeometryVectors &powerlineNetwork) {
-                m_impl->addVectorToTilesetCollection(powerlineNetwork,
-                                                     powerlineNetworkDir,
-                                                     m_impl->powerlineNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->powerlineNetworkTilesets);
-
-            cdb.forEachHydrographyNetworkTile(geoCell, [&](const CDBGeometryVectors &hydrographyNetwork) {
-                m_impl->addVectorToTilesetCollection(hydrographyNetwork,
-                                                     hydrographyNetworkDir,
-                                                     m_impl->hydrographyNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->hydrographyNetworkTilesets);
-
-            std::set<std::string> subtreeRoots;
-            std::map<CDBDataset, std::map<std::string,std::map<std::string, SubtreeAvailability>>>
-                &datasetCSTileAndChildAvailabilities = m_impl->datasetCSTileAndChildAvailabilities;
-
-            // write all of the availability buffers and subtree files for each dataset group
-            for (auto &[dataset, csTileAndChildAvailabilities] : datasetCSTileAndChildAvailabilities) {
-                if (datasetCSSubtrees.count(dataset) == 0) {
-                    continue;
-                }
-                for (auto &[CSKey, subtreeMap] : datasetCSSubtrees.at(dataset)) {
-                    std::map<std::string, SubtreeAvailability> &tileAndChildAvailabilities = 
-                        csTileAndChildAvailabilities.at(CSKey);
-                    for (auto &[key, subtree] : subtreeMap) {
-                        subtreeRoots.insert(key);
-
-                        bool constantNodeAvailability = (subtree.nodeCount == 0)
-                                                        || (subtree.nodeCount == subtreeNodeCount);
-
-                        if (constantNodeAvailability) {
-                            continue;
-                        }
-
-                        std::vector<uint8_t> outputBuffer(nodeAvailabilityByteLengthWithPadding);
-                        uint8_t *outBuffer = &outputBuffer[0];
-                        memset(&outBuffer[0], 0, nodeAvailabilityByteLengthWithPadding);
-                        memcpy(&outBuffer[0],
-                                &subtree.nodeBuffer[0],
-                                nodeAvailabilityByteLengthWithPadding);
-                        std::filesystem::path path = datasetDirs.at(dataset) / CSKey / "availability"
-                                                        / (key + ".bin");
-                        Utilities::writeBinaryFile(path,
-                                                    (const char *) &outBuffer[0],
-                                                    nodeAvailabilityByteLengthWithPadding);
-                    }
-
-                    // write .subtree files for every subtree
-                    for (std::string subtreeRoot : subtreeRoots) {
-                        json subtreeJson;
-
-                        nlohmann::json buffers = nlohmann::json::array();
-                        int bufferIndex = 0;
-                        nlohmann::json bufferViews = nlohmann::json::array();
-                        SubtreeAvailability tileAndChildAvailability = tileAndChildAvailabilities.at(subtreeRoot);
-                        tileAndChildAvailability.nodeCount = countSetBitsInVectorOfInts(
-                            tileAndChildAvailability.nodeBuffer);
-                        tileAndChildAvailability.childCount = countSetBitsInVectorOfInts(
-                            tileAndChildAvailability.childBuffer);
-                        bool constantTileAvailability = (tileAndChildAvailability.nodeCount == 0)
-                                                        || (tileAndChildAvailability.nodeCount
-                                                            == subtreeNodeCount);
-                        bool constantChildAvailability = (tileAndChildAvailability.childCount == 0)
-                                                        || (tileAndChildAvailability.childCount
-                                                            == childSubtreeCount);
-
-                        uint64_t nodeBufferLengthToWrite = static_cast<int>(!constantTileAvailability)
-                                                        * nodeAvailabilityByteLengthWithPadding;
-                        uint64_t childBufferLengthToWrite = static_cast<int>(!constantChildAvailability)
-                                                            * childSubtreeAvailabilityByteLengthWithPadding;
-                        long unsigned int bufferByteLength = nodeBufferLengthToWrite + childBufferLengthToWrite;
-                        if (bufferByteLength != 0) {
-                            nlohmann::json byteLength;
-                            byteLength["byteLength"] = bufferByteLength;
-                            buffers.emplace_back(byteLength);
-                            bufferIndex += 1;
-                        }
-
-                        std::vector<uint8_t> internalBuffer(bufferByteLength);
-                        memset(&internalBuffer[0], 0, bufferByteLength);
-                        uint8_t *outInternalBuffer = &internalBuffer[0];
-                        nlohmann::json tileAvailabilityJson;
-                        int bufferViewIndex = 0;
-                        uint64_t internalBufferOffset = 0;
-                        if (constantTileAvailability)
-                            tileAvailabilityJson["constant"] = static_cast<int>(tileAndChildAvailability.nodeCount
-                                                                                == subtreeNodeCount);
-                        else {
-                            memcpy(&outInternalBuffer[0],
-                                &tileAndChildAvailability.nodeBuffer[0],
-                                nodeAvailabilityByteLengthWithPadding);
-                            nlohmann::json bufferViewObj;
-                            bufferViewObj["buffer"] = 0;
-                            bufferViewObj["byteOffset"] = 0;
-                            bufferViewObj["byteLength"] = availabilityByteLength;
-                            bufferViews.emplace_back(bufferViewObj);
-                            internalBufferOffset += nodeAvailabilityByteLengthWithPadding;
-                            tileAvailabilityJson["bufferView"] = bufferViewIndex;
-                            bufferViewIndex += 1;
-                        }
-                        subtreeJson["tileAvailability"] = tileAvailabilityJson;
-
-                        nlohmann::json childAvailabilityJson;
-                        if (constantChildAvailability)
-                            childAvailabilityJson["constant"] = static_cast<int>(
-                                tileAndChildAvailability.childCount == childSubtreeCount);
-                        else {
-                            memcpy(&outInternalBuffer[internalBufferOffset],
-                                &tileAndChildAvailability.childBuffer[0],
-                                childSubtreeAvailabilityByteLengthWithPadding);
-                            nlohmann::json bufferViewObj;
-                            bufferViewObj["buffer"] = 0;
-                            bufferViewObj["byteOffset"] = internalBufferOffset;
-                            bufferViewObj["byteLength"] = childSubtreeAvailabilityByteLength;
-                            bufferViews.emplace_back(bufferViewObj);
-                            childAvailabilityJson["bufferView"] = bufferViewIndex;
-                            bufferViewIndex += 1;
-                        }
-                        subtreeJson["childSubtreeAvailability"] = childAvailabilityJson;
-
-                        nlohmann::json contentAvailability = nlohmann::json::array();
-                        std::string availabilityFileName = subtreeRoot + ".bin";
-
-                        std::filesystem::path datasetDir = datasetDirs.at(dataset);
-
-                        std::map<std::string, SubtreeAvailability> csSubtreeRoots
-                            = datasetCSSubtrees.at(dataset).at(CSKey);
-                        nlohmann::json contentObj;
-                        if (std::filesystem::exists(datasetDir / CSKey / "availability"
-                                                    / availabilityFileName)) {
-                            nlohmann::json bufferObj;
-                            auto datasetDirIt = datasetDir.end();
-                            --datasetDirIt; // point to the dataset directory name
-                            bufferObj["uri"] = "../availability/" + availabilityFileName;
-                            bufferObj["byteLength"] = nodeAvailabilityByteLengthWithPadding;
-                            buffers.emplace_back(bufferObj);
-                            nlohmann::json bufferViewObj;
-                            bufferViewObj["buffer"] = bufferIndex;
-                            bufferViewObj["byteOffset"] = 0;
-                            bufferViewObj["byteLength"] = availabilityByteLength;
-                            bufferViews.emplace_back(bufferViewObj);
-                            contentObj["bufferView"] = bufferViewIndex;
-                            bufferViewIndex += 1;
-                            bufferIndex += 1;
-                        } else if (csSubtreeRoots.count(subtreeRoot) != 0) {
-                            SubtreeAvailability subtree = datasetCSSubtrees.at(dataset).at(CSKey).at(
-                                subtreeRoot);
-                            contentObj["constant"] = static_cast<int>(subtree.nodeCount == subtreeNodeCount);
-                        } else
-                            contentObj["constant"] = 0;
-                        contentAvailability.emplace_back(contentObj);
-                        subtreeJson["contentAvailability"] = contentAvailability;
-                        if (!buffers.empty())
-                            subtreeJson["buffers"] = buffers;
-                        if (!bufferViews.empty())
-                            subtreeJson["bufferViews"] = bufferViews;
-
-                        // get json length
-                        const std::string jsonString = subtreeJson.dump();
-                        const uint64_t jsonStringByteLength = jsonString.size();
-                        const uint64_t jsonStringByteLengthWithPadding = alignTo8(jsonStringByteLength);
-
-                        // Write subtree binary
-                        uint64_t outputBufferLength = jsonStringByteLengthWithPadding + bufferByteLength
-                                                    + headerByteLength;
-                        std::vector<uint8_t> outputBuffer(outputBufferLength);
-                        uint8_t *outBuffer = &outputBuffer[0];
-                        *(uint32_t *) &outBuffer[0] = 0x74627573; // magic: "subt"
-                        *(uint32_t *) &outBuffer[4] = 1;          // version
-                        *(uint64_t *) &outBuffer[8]
-                            = jsonStringByteLengthWithPadding;           // JSON byte length with padding
-                        *(uint64_t *) &outBuffer[16] = bufferByteLength; // BIN byte length with padding
-
-                        memcpy(&outBuffer[headerByteLength], &jsonString[0], jsonStringByteLength);
-                        memset(&outBuffer[headerByteLength + jsonStringByteLength],
-                            ' ',
-                            jsonStringByteLengthWithPadding - jsonStringByteLength);
-
-                        if (bufferByteLength != 0) {
-                            memcpy(&outBuffer[headerByteLength + jsonStringByteLengthWithPadding],
-                                outInternalBuffer,
-                                bufferByteLength);
-                        }
-                        std::filesystem::path path = datasetDir / CSKey / "subtrees"
-                                                    / (subtreeRoot + ".subtree");
-                        Utilities::writeBinaryFile(path, (const char *) outBuffer, outputBufferLength);
-                    }
-                    tileAndChildAvailabilities.clear();
-                    subtreeRoots.clear();
-                }
-            }
-
-            // get the converted dataset in each geocell to be combine at the end
-            Core::BoundingRegion geoCellRegion = CDBTile::calcBoundRegion(geoCell, -10, 0, 0);
-            for (auto tilesetJsonPath : m_impl->defaultDatasetToCombine) {
-                auto componentSelectors = tilesetJsonPath.parent_path().filename().string();
-                auto dataset = tilesetJsonPath.parent_path().parent_path().filename().string();
-                auto combinedTilesetName = dataset + "_" + componentSelectors;
-
-                combinedTilesets[combinedTilesetName].emplace_back(tilesetJsonPath);
-                combinedTilesetsRegions[combinedTilesetName].emplace_back(geoCellRegion);
-                auto tilesetAggregateRegion = aggregateTilesetsRegion.find(combinedTilesetName);
-                if (tilesetAggregateRegion == aggregateTilesetsRegion.end()) {
-                    aggregateTilesetsRegion.insert({combinedTilesetName, geoCellRegion});
-                } else {
-                    tilesetAggregateRegion->second = tilesetAggregateRegion->second.computeUnion(
-                        geoCellRegion);
-                }
-            }
-            std::vector<std::filesystem::path>().swap(m_impl->defaultDatasetToCombine);
+        // process elevation
+        cdb.forEachElevationTile(geoCell, [&](CDBElevation elevation) {
+            m_impl->addElevationToTilesetCollection(elevation, cdb, elevationDir);
         });
+        m_impl->flushTilesetCollection(geoCell, m_impl->elevationTilesets);
+        std::unordered_map<CDBTile, Texture>().swap(m_impl->processedParentImagery);
 
-        // combine all the default tileset in each geocell into a global one
-        for (auto tileset : combinedTilesets) {
-            std::ofstream fs(m_impl->outputPath / (tileset.first + ".json"));
-            combineTilesetJson(tileset.second, combinedTilesetsRegions[tileset.first], fs);
-        }
-
-        // combine the requested tilesets
-        for (const auto &tilesets : m_impl->requestedDatasetToCombine) {
-            std::string combinedTilesetName;
-            if (m_impl->requestedDatasetToCombine.size() > 1) {
-                for (const auto &tileset : tilesets) {
-                    combinedTilesetName += tileset;
-                }
-                combinedTilesetName += ".json";
-            } else {
-                combinedTilesetName = "tileset.json";
-            }
-
-            std::vector<std::filesystem::path> existTilesets;
-            std::vector<Core::BoundingRegion> regions;
-            regions.reserve(tilesets.size());
-            for (const auto &tileset : tilesets) {
-                auto tilesetRegion = aggregateTilesetsRegion.find(tileset);
-                if (tilesetRegion != aggregateTilesetsRegion.end()) {
-                    existTilesets.emplace_back(tilesetRegion->first + ".json");
-                    regions.emplace_back(tilesetRegion->second);
-                }
-            }
-
-            std::ofstream fs(m_impl->outputPath / combinedTilesetName);
-            combineTilesetJson(existTilesets, regions, fs);
-        }
-    } else {
-        cdb.forEachGeoCell([&](CDBGeoCell geoCell) {
-            // create directories for converted GeoCell
-            std::filesystem::path geoCellRelativePath = geoCell.getRelativePath();
-            std::filesystem::path geoCellAbsolutePath = m_impl->outputPath / geoCellRelativePath;
-            std::filesystem::path elevationDir = geoCellAbsolutePath / CDBTilesetBuilder::ELEVATIONS_PATH;
-            std::filesystem::path GTModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GTMODEL_PATH;
-            std::filesystem::path GSModelDir = geoCellAbsolutePath / CDBTilesetBuilder::GSMODEL_PATH;
-            std::filesystem::path roadNetworkDir = geoCellAbsolutePath / CDBTilesetBuilder::ROAD_NETWORK_PATH;
-            std::filesystem::path railRoadNetworkDir = geoCellAbsolutePath
-                                                       / CDBTilesetBuilder::RAILROAD_NETWORK_PATH;
-            std::filesystem::path powerlineNetworkDir = geoCellAbsolutePath
-                                                        / CDBTilesetBuilder::POWERLINE_NETWORK_PATH;
-            std::filesystem::path hydrographyNetworkDir = geoCellAbsolutePath
-                                                          / CDBTilesetBuilder::HYDROGRAPHY_NETWORK_PATH;
-
-            // process elevation
-            cdb.forEachElevationTile(geoCell, [&](CDBElevation elevation) {
-                m_impl->addElevationToTilesetCollection(elevation, cdb, elevationDir);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->elevationTilesets);
-            std::unordered_map<CDBTile, Texture>().swap(m_impl->processedParentImagery);
-
-            // process road network
-            cdb.forEachRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &roadNetwork) {
-                m_impl->addVectorToTilesetCollection(roadNetwork, roadNetworkDir, m_impl->roadNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->roadNetworkTilesets);
-
-            // process railroad network
-            cdb.forEachRailRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &railRoadNetwork) {
-                m_impl->addVectorToTilesetCollection(railRoadNetwork,
-                                                     railRoadNetworkDir,
-                                                     m_impl->railRoadNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->railRoadNetworkTilesets);
-
-            // process powerline network
-            cdb.forEachPowerlineNetworkTile(geoCell, [&](const CDBGeometryVectors &powerlineNetwork) {
-                m_impl->addVectorToTilesetCollection(powerlineNetwork,
-                                                     powerlineNetworkDir,
-                                                     m_impl->powerlineNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->powerlineNetworkTilesets);
-
-            // process hydrography network
-            cdb.forEachHydrographyNetworkTile(geoCell, [&](const CDBGeometryVectors &hydrographyNetwork) {
-                m_impl->addVectorToTilesetCollection(hydrographyNetwork,
-                                                     hydrographyNetworkDir,
-                                                     m_impl->hydrographyNetworkTilesets);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->hydrographyNetworkTilesets);
-
-            // process GTModel
-            // TODO: Remove this workaround when EXT_mesh_gpu_instancing is implemented
-            if (!m_impl->use3dTilesNext) {
-                cdb.forEachGTModelTile(geoCell, [&](CDBGTModels GTModel) {
-                    m_impl->addGTModelToTilesetCollection(GTModel, GTModelDir);
-                });
-                m_impl->flushTilesetCollection(geoCell, m_impl->GTModelTilesets);
-            }
-
-            // process GSModel
-            cdb.forEachGSModelTile(geoCell, [&](CDBGSModels GSModel) {
-                m_impl->addGSModelToTilesetCollection(GSModel, GSModelDir);
-            });
-            m_impl->flushTilesetCollection(geoCell, m_impl->GSModelTilesets, false);
-
-            // get the converted dataset in each geocell to be combine at the end
-            Core::BoundingRegion geoCellRegion = CDBTile::calcBoundRegion(geoCell, -10, 0, 0);
-            for (auto tilesetJsonPath : m_impl->defaultDatasetToCombine) {
-                auto componentSelectors = tilesetJsonPath.parent_path().filename().string();
-                auto dataset = tilesetJsonPath.parent_path().parent_path().filename().string();
-                auto combinedTilesetName = dataset + "_" + componentSelectors;
-
-                combinedTilesets[combinedTilesetName].emplace_back(tilesetJsonPath);
-                combinedTilesetsRegions[combinedTilesetName].emplace_back(geoCellRegion);
-                auto tilesetAggregateRegion = aggregateTilesetsRegion.find(combinedTilesetName);
-                if (tilesetAggregateRegion == aggregateTilesetsRegion.end()) {
-                    aggregateTilesetsRegion.insert({combinedTilesetName, geoCellRegion});
-                } else {
-                    tilesetAggregateRegion->second = tilesetAggregateRegion->second.computeUnion(
-                        geoCellRegion);
-                }
-            }
-            std::vector<std::filesystem::path>().swap(m_impl->defaultDatasetToCombine);
+        // process road network
+        cdb.forEachRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &roadNetwork) {
+            m_impl->addVectorToTilesetCollection(roadNetwork, roadNetworkDir, m_impl->roadNetworkTilesets);
         });
+        m_impl->flushTilesetCollection(geoCell, m_impl->roadNetworkTilesets);
 
-        // combine all the default tileset in each geocell into a global one
-        for (auto tileset : combinedTilesets) {
-            std::ofstream fs(m_impl->outputPath / (tileset.first + ".json"));
-            combineTilesetJson(tileset.second, combinedTilesetsRegions[tileset.first], fs);
-        }
+        // process railroad network
+        cdb.forEachRailRoadNetworkTile(geoCell, [&](const CDBGeometryVectors &railRoadNetwork) {
+            m_impl->addVectorToTilesetCollection(railRoadNetwork,
+                                                    railRoadNetworkDir,
+                                                    m_impl->railRoadNetworkTilesets);
+        });
+        m_impl->flushTilesetCollection(geoCell, m_impl->railRoadNetworkTilesets);
 
-        // combine the requested tilesets
-        for (const auto &tilesets : m_impl->requestedDatasetToCombine) {
-            std::string combinedTilesetName;
-            if (m_impl->requestedDatasetToCombine.size() > 1) {
-                for (const auto &tileset : tilesets) {
-                    combinedTilesetName += tileset;
-                }
-                combinedTilesetName += ".json";
+        // process powerline network
+        cdb.forEachPowerlineNetworkTile(geoCell, [&](const CDBGeometryVectors &powerlineNetwork) {
+            m_impl->addVectorToTilesetCollection(powerlineNetwork,
+                                                    powerlineNetworkDir,
+                                                    m_impl->powerlineNetworkTilesets);
+        });
+        m_impl->flushTilesetCollection(geoCell, m_impl->powerlineNetworkTilesets);
+
+        // process hydrography network
+        cdb.forEachHydrographyNetworkTile(geoCell, [&](const CDBGeometryVectors &hydrographyNetwork) {
+            m_impl->addVectorToTilesetCollection(hydrographyNetwork,
+                                                    hydrographyNetworkDir,
+                                                    m_impl->hydrographyNetworkTilesets);
+        });
+        m_impl->flushTilesetCollection(geoCell, m_impl->hydrographyNetworkTilesets);
+
+        // process GTModel
+        cdb.forEachGTModelTile(geoCell, [&](CDBGTModels GTModel) {
+            m_impl->addGTModelToTilesetCollection(GTModel, GTModelDir);
+        });
+        m_impl->flushTilesetCollection(geoCell, m_impl->GTModelTilesets);
+
+        // process GSModel
+        cdb.forEachGSModelTile(geoCell, [&](CDBGSModels GSModel) {
+            m_impl->addGSModelToTilesetCollection(GSModel, GSModelDir);
+        });
+        m_impl->flushTilesetCollection(geoCell, m_impl->GSModelTilesets, false);
+
+        m_impl->flushAvailabilitiesAndWriteSubtrees();
+
+        // get the converted dataset in each geocell to be combine at the end
+        Core::BoundingRegion geoCellRegion = CDBTile::calcBoundRegion(geoCell, -10, 0, 0);
+        for (auto tilesetJsonPath : m_impl->defaultDatasetToCombine) {
+            auto componentSelectors = tilesetJsonPath.parent_path().filename().string();
+            auto dataset = tilesetJsonPath.parent_path().parent_path().filename().string();
+            auto combinedTilesetName = dataset + "_" + componentSelectors;
+
+            combinedTilesets[combinedTilesetName].emplace_back(tilesetJsonPath);
+            combinedTilesetsRegions[combinedTilesetName].emplace_back(geoCellRegion);
+            auto tilesetAggregateRegion = aggregateTilesetsRegion.find(combinedTilesetName);
+            if (tilesetAggregateRegion == aggregateTilesetsRegion.end()) {
+                aggregateTilesetsRegion.insert({combinedTilesetName, geoCellRegion});
             } else {
-                combinedTilesetName = "tileset.json";
+                tilesetAggregateRegion->second = tilesetAggregateRegion->second.computeUnion(
+                    geoCellRegion);
             }
-
-            std::vector<std::filesystem::path> existTilesets;
-            std::vector<Core::BoundingRegion> regions;
-            regions.reserve(tilesets.size());
-            for (const auto &tileset : tilesets) {
-                auto tilesetRegion = aggregateTilesetsRegion.find(tileset);
-                if (tilesetRegion != aggregateTilesetsRegion.end()) {
-                    existTilesets.emplace_back(tilesetRegion->first + ".json");
-                    regions.emplace_back(tilesetRegion->second);
-                }
-            }
-
-            std::ofstream fs(m_impl->outputPath / combinedTilesetName);
-            combineTilesetJson(existTilesets, regions, fs);
         }
+        std::vector<std::filesystem::path>().swap(m_impl->defaultDatasetToCombine);
+    });
+
+    // combine all the default tileset in each geocell into a global one
+    for (auto tileset : combinedTilesets) {
+        std::ofstream fs(m_impl->outputPath / (tileset.first + ".json"));
+        combineTilesetJson(tileset.second, combinedTilesetsRegions[tileset.first], fs);
+    }
+
+    // combine the requested tilesets
+    for (const auto &tilesets : m_impl->requestedDatasetToCombine) {
+        std::string combinedTilesetName;
+        if (m_impl->requestedDatasetToCombine.size() > 1) {
+            for (const auto &tileset : tilesets) {
+                combinedTilesetName += tileset;
+            }
+            combinedTilesetName += ".json";
+        } else {
+            combinedTilesetName = "tileset.json";
+        }
+
+        std::vector<std::filesystem::path> existTilesets;
+        std::vector<Core::BoundingRegion> regions;
+        regions.reserve(tilesets.size());
+        for (const auto &tileset : tilesets) {
+            auto tilesetRegion = aggregateTilesetsRegion.find(tileset);
+            if (tilesetRegion != aggregateTilesetsRegion.end()) {
+                existTilesets.emplace_back(tilesetRegion->first + ".json");
+                regions.emplace_back(tilesetRegion->second);
+            }
+        }
+
+        std::ofstream fs(m_impl->outputPath / combinedTilesetName);
+        combineTilesetJson(existTilesets, regions, fs);
     }
 }
